jdk版本升级17 spring版本升级2.7.x

+++ src/main/java/com/ShiXi/team/service/impl/TeamServiceImpl.java
@@ -531,7 +531,7 @@ public class TeamServiceImpl extends ServiceImpl<TeamMapper, Team> implements Te
         }
 
         // 获取团队成员数量
-        Integer memberCount = teamMemberService.lambdaQuery()
+        Long memberCount = teamMemberService.lambdaQuery()
                 .eq(TeamMember::getTeamId, team.getId())
                 .eq(TeamMember::getStatus, "approved")
                 .count();
